Remove hard dependency on the common/license file for license detection.
This is to allow users to use autospec as a standalone package.
If the license server is not provided, then the common/license file is always read.
Removing this dependency, will allow the license to be picked by other means
($package.license file). Only when license detection is not possible through any of the
methods, an error is flagged.