Generate an ABI report for dynamic symbols through a new ./symbols file
This file follows the format of "$soname:$symbol", and will be exported
in the instance that the package being autospecced has successfully built
binary .rpm's, and contains shared libraries with exported symbols.

Each $soname and $symbol is sorted, so successive builds of a package should
only ever create an alteration to the symbols file in the instance that symbols
truly changed, enabling stronger CI practices, by way of preempting any
potential ABI breaks.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>