abireport: Collect directly used dependencies in the used_libs output file
Much as abireport gave us the symbols output file in order to know exactly
what the ABI looks like at a given revision, we now add support to collect
all of the _direct_ binary dependencies at build time into `used_libs`.

This file is always sorted in order to ensure there is never a diff in the
git commit unless something genuinely changed. Likewise, internally provided
symbols are excluded from the generated list, such as when packages provide
their own versions of a library that may be present on the system with the
same soname.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>