Improve configure.ac parsing of nested constructs
When scanning the configure.ac for build requirements, the file is first
parsed into "blocks", where each block might be a simple Autoconf macro
call or a nested construct with more than one macro call.

However, when parsing the concatenated string that represents a nested
construct, the regular expression used to detect the macro calls use the
greedy '*' metacharacter, which captures too much of the line (more than
a single macro call). Since the intention is to capture just the single
macro call, use the non-greedy form '*?' instead.

Note that this solution covers the common situation, where quoted values
are passed to PKG_CHECK_MODULES and similar macros. If a macro is passed
as the value instead, then a different implementation is needed in order
to perform full macro expansion first, and then extract the appropriate
value. One idea is use 'autoconf --trace=MACRO_NAME' to parse macro
values, but this solution seems to not work with third-party
(non-Autoconf) macros.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>