Use pycurl instead of urllib.
urllib does not use any proxy smartness, causing problems for the
two uses of urllib in autospec. You may end up having to disable
proxies to make the first step (license server interaction) work,
and then re-enable it for the tarball download.

This patch delegates everything to libcurl.so, which knows about
proxies and does the right thing.