Adding quotes to the filenames in file section of specfile
There are some cases when filenames are added to the file section in the specfile for checking.
If the files contains spaces, rpmbuild will crash due to parsing it as multiple files.
This patch adds quotation to the filename whenever a space is found to avoid rpmbuild
to crash in file checking.

Signed-off-by: Gabriel Briones Sayeg <gabriel.briones.sayeg@intel.com>