Add support for zip files with hashes
When zip files contain hash information it is displayed in the header
causing failure to detect the zip file prefix. This change adds minimal
zip file header detection in order to determine which line contains the
first compressed file to find the prefix.

An example of the unzip -l output that had previously been the only
correctly handled type is:

Archive:  wsgiref-0.1.2.zip
  Length      Date    Time    Name
---------  ---------- -----   ----
      589  06-11-2006 23:49   wsgiref-0.1.2/PKG-INFO
      239  06-07-2006 13:13   wsgiref-0.1.2/README.txt
       51  06-07-2006 13:10   wsgiref-0.1.2/setup.cfg
      581  06-11-2006 23:48   wsgiref-0.1.2/setup.py
    17925  06-11-2006 23:18   wsgiref-0.1.2/test_wsgiref.py
    27441  06-09-2006 11:54   wsgiref-0.1.2/docs/libwsgiref.tex
     2295  06-09-2006 12:01   wsgiref-0.1.2/docs/ref.tex
      614  07-09-2005 00:29   wsgiref-0.1.2/ez_setup/README.txt
     7661  06-09-2006 14:43   wsgiref-0.1.2/ez_setup/__init__.py
    15596  06-11-2006 23:44   wsgiref-0.1.2/wsgiref/handlers.py
     5916  06-11-2006 23:43   wsgiref-0.1.2/wsgiref/headers.py
     4789  06-11-2006 23:43   wsgiref-0.1.2/wsgiref/simple_server.py
     5598  06-11-2006 23:44   wsgiref-0.1.2/wsgiref/util.py
    14737  06-11-2006 23:42   wsgiref-0.1.2/wsgiref/validate.py
      586  06-05-2006 13:45   wsgiref-0.1.2/wsgiref/__init__.py
      589  06-11-2006 23:49   wsgiref-0.1.2/wsgiref.egg-info/PKG-INFO
      343  06-11-2006 23:49   wsgiref-0.1.2/wsgiref.egg-info/SOURCES.txt
        8  06-11-2006 23:49   wsgiref-0.1.2/wsgiref.egg-info/top_level.txt
---------                     -------
   105558                     18 files

The first line after the header:

      589  06-11-2006 23:49   wsgiref-0.1.2/PKG-INFO

is on the 4th line of the output in this case unlike the output where a
checksum is embedded in the zip file:

Archive:  msr-tools-1.3.zip
d9a630da8b626196737d29dff697bfda405de93e
  Length      Date    Time    Name
---------  ---------- -----   ----
        0  09-10-2013 02:49   msr-tools-master/
      231  09-10-2013 02:49   msr-tools-master/MAKEDEV-cpuid-msr
      977  09-10-2013 02:49   msr-tools-master/Makefile
     4383  09-10-2013 02:49   msr-tools-master/cpuid.c
     7645  09-10-2013 02:49   msr-tools-master/rdmsr.c
      102  09-10-2013 02:49   msr-tools-master/version.h
     3756  09-10-2013 02:49   msr-tools-master/wrmsr.c
---------                     -------
    17094                     7 files

In this case the first line after the header:

        0  09-10-2013 02:49   msr-tools-master/

is on the 5th line of the output causing the header separation line to
be erroneously used to get the prefix.

In addition, refactor the code into a function since the unzip logic was
used in multiple places already.