Add autostart subpackage
This change allows packages to have an autostart subpackage that is
included by default as one of the main package's dependencies. It is
intended to replace an informal use of the extras subpackage which
previously handled having a separation between systemd unit activation
files and the main package.

To make use of this feature, a packager will likely create the
target.wants folder for the systemd unit and create the symlink to the
systemd unit in make_install_append and add the full path to the
unit file symlink in the autostart file (as described in the
README.rst).