Rework autostart implementation
This change prevents the possibility of enabled systemd unit files not
being put into the autostart subpackage by just matching all files in
the system *.target.wants subdirectories. The autostart config file has
thus been reworked to instead be a no_autostart file that controls if
the autostart subpackage is included by the main package or not.