Add basic functional test framework to Autospec
Add test framework and basic tests to Autospec to allow testing of
packages. The tests can be run by invoking `python3 test_autospec.py`

The tests first create a tarball of each source directory within the
testfiles directory and runs autospec against it. Each package is
handled in its own thread, so the total time of the tests is equivalent
to the total time of the longest autospec run.

Once the build is complete, the output of `make autospec` and the
command resulting spec file are searched for terms defined by in a
special file within the test source directories. The special file is
expectations.py and within this file the following terms can be
defined:

- buildreqs (list of strings) - build requirements expected to be
detected by autospec.

- license (string) - name of the license to be detected by autospec.

- output_strings (list of strings) - strings expected to be in the
output of `make autospec`.

- specfile (string) - this string is read in from the spec-expectations
file, which contains the exact text of the specfile expected to be
generated from the autospec run.

The file expectations.py is not included in the tarball created in the
first step of the test process.

There is an additional directory 'autospecdir' within the test package
that contains the configuration files used by an autospec run.

Test results are printed at the end of an autospec run on a package. If
the generated specfile differs from spec-expectations, a diff is
displayed along with the failure.

tarball.py is also updated to remove only the specific tarball that was
downloaded. This allows the tests to be run in parallel without the
"downloaded" tarball being erased by subsequent autospec runs.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>