Rework golang support
The support for golang in autospec was previously installing packages
into the GOROOT causing failures when go get was run on packages already
in the GOROOT.

This patch moves the install location to /usr/lib/golang-dist which
should be added to the GOPATH when building a go package.

The autospec support still does not allow for building go and instead
just handles adding library installation at this time. It also requires
manual configuration for the go source path and build pattern.