Rewrite options configuration file on autospec run
options.conf may exist in a package directory in a format different from
the format defined in config.py due to an out-of-date autospec or
user-creation of the configuration file. Rewriting the file on each run
updates the configuration file to the common format without requiring
manual editing by the user.

Consequently, this also allows the user to provide a partial
configuration file without defining every option required. The
configuration file will be rewritten with all available options and
associated comments. Previously missing options will be added as false
values.