Remove downloaded archives at beginning of run
Archives are downloaded then moved to /tmp/<user>/<tarball_prefix>/
If this directory is not removed before an autospec run and a user runs
autospec on the same package twice in a row, the build will fail due to
the archives existing in the same location the new archives are
attempting to be moved to. This directory needs to be removed before an
autospec run to prevent this.