Abstracting url signature extension
Some URLs conform to a pttern for the package signature
extension. To handle this pattern, the decision of the
type of extension for the signature URL was moved to
under a function. Test contains a fixed needed after
the extension abstraction that helped with validation
of packages from certain sources that were not easy
to validate before.