Separate out name and version detection
Presently the package name and version is determined from the tarball
URL in the tarball.download_tarball() function. This is not a logical
place to do the detection, nor does it make name and version detection
easy to test.

Separating it out into its own function makes logical sense and allows
unit testing of the name and version detection.