Consolidate spec file formation code
This patch consolidates almost all code that writes to the spec file
into one place.

This is beneficial for a number of reasons. All the write logic is
co-located in the specfile.py file and therefore easier to debug. This
file is the core functionality of autospec, so it makes sense to keep it
all in one place instead of spread out over eight different modules.

The Specfile class is written in such a way that once all the instance
variables are populated it will write a valid spec file. This is a major
advantage because it allows for more deterministic spec file creation
regardless of changes elsewhere in the codebase.

The current exception to the above rules is that Specfile relies on
buildpattern.py methods for its write_buildpattern method. This is
because buildpattern.py uses pattern matching to dynamically run a
module method based on the buildpattern of the current package. This
should be fixed in the future. There is also a small dependency on
config.py for writing the scriplets.