Move buildpattern writes to Specfile class
The buildpattern module was an extension of the specfile write logic,
living in its own place. This patch moves all remaining specfile write
logic to one place - the Specfile class. The Specfile class is now
dependent on only one other module, the config module. With this change
the Specfile class can act as documentation of our spec file policy.

The API for the Specfile class is somewhat simpler than before because
no writes to the spec file are done outside of the class. One downside
of doing it this way is the large number of instance variables that need
to be set before the writing can take place. This is still an
improvement over the previous implementation, however, because now all
that data is in the same place.

Two modules, lang and patches, were removed by being included in other
modules once their spec file write code was moved to the Specfile class.
The lang module was moved to the files module and patches was moved into
config. The buildpattern module is now very simple but remains to hold
several global variables that are deeply integrated into code elsewhere.