Always use distutils23
Previous settings would largely force distutils (python2 only) builds
for python source. This was done in tarball.py as it would set the build
pattern based on the url of the source. This change updates that setting
to use distutils23 instead.

This change also removes rather broken python3 detection in setup.py or
patch files.