Add cargo support to autospec
This change adds the ability for autospec to package projects using the
cargo build tool. Note that source files are installed into
/usr/share/rust/src and an empty checksums file is prepopulated into the
source folders for now. This is an implementation taking advantage of
cargo's recently added ability to handle vendored source directories and
may evolve further as distro support is still not fully baked in.