Validate gpg signature before using it
Because some repos return 200 http code with a
404 page when a signature is not found, validating
signatures is needed to make sure the downloaded
data can be read by gpg.

This will print the proper error messages for
CLEAR-1752 and CLEAR-1756, right now it just
returns a None and gives the impression that
a signature is available, which is missleading.