abireport: Add support for host-side abireport tool
The abireport portion of autospec has since been ported (in concept) to
a new Golang tool of the same name. This tool is decidedly more accurate
and considerably faster, as it does not have to call out to execute more
system commands for each file processed.

In addition, it is able to differentiate correctly between multiple
architectures, allowing for two new files to exist, used_libs32 and
symbols32. These permit a finer granularity in the view of multilib
as exposed by the package.

In the instance that abireport is not available on the host, we fall
back to the internal abireport scanning, and emit a warning.

Minor tweaks have been made to the binary_in_path function to make it
reusable as part of util, and to make it lazily initialise the os paths
to cut down on unnecessary duplication.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>