Simplify build.py pattern matching
Move large configuration lists to config.py and combine the
failed_pattern* methods into one. This reduces code repetition and
simplifies the logic of the build.log parser.

It is normally a goal to move configuration lists to a separate text
file. These lists contain python regex patterns, so they should stay in
a python file to ensure correct syntax.