Completely remove results directory on new run
Instead of allowing a stale results directory to hang around during a
new autospec run, remove it completely before the mock build. If it is
left there untouched, old files and rpms may not be overwritten by the
new build. This has been happening with the new abireport written in go
being included in autospec - a new rpm is created and an old one is left
untouched in the results directory.