Ensure required packages exist before including
In order to avoid making spec files that add invalid packages to
requires lists, compare the requirement to the buildrequires list or the
list of os packages first.

There is a slightly unpleasant workaround being done for handling
relative paths when a looking for the common/packages file so I'm open
to other ways of handling this without needing to have a copy of the
packages file for each person on the same system using the default
setup (projects/{autospec,common} existing).