Convert test_autospec output to utf-8
When running build_and_run for test_autospec if subprocess.check_output
throws an exception, output was a byte encoded string and a later test
to see if a unicode string was contained in that byte encoded string
would throw an exception.

Instead of doing the decode in a try block put it in finally. This does
add the posibility of utf-8 decode failing but that's fatal later on
either way.