Only grab prefix from first tarball contents directory
Instead of grabbing the tarball prefix from the first or second line in
the tarball contents, iterate through the lines until a properly formed
line is found. The prefix will usually be grabbed from the first line,
but cases where we have to keep looking are as follows:

tar -tf <tarfile>

name-with-no-trailing-slash
name-with-trailing-slash/README
name-with-trailing-slash/...
...

or

./
./name-with-trailing-slash/README
./name-with-trailing-slash/...
...

In both cases we want "name-with-trailing-slash." If the tarball prefix
cannot be determined, a fatal error will be printed and autospec will
exit.