Rework go support in autospec
This change allows autospec to build and ship go binaries but stops
autospec creating source directories for the packages. This means that
autospec is to be used only to create go binaries and thus those
binaries must vendor all their dependencies.

Go packages will likely need to be patched in order for go build to
work due to the directory naming and/or vendoring handling. Please see
the patch in the 0001 patch in Clear Linux's
golang-github-cpuguy83-go-md2man package for an example of this.