Port logcheck to python and add to autospec
The perl logcheck tool is helpful in tracking package changes between
versions. This patch ports the script to python and includes it in the
autospec flow on successful builds. Two more configuration files
(configure_whitelist and configure_blacklist) are added to autospec as
well. The results are logged to a "configure_misses" file in the package
directory and checked into git.

This will be very helpful for other tools that attempt to
programmatically determine the confidence level of a package update.