Fix inadequate python build requirement parsing
Build requirement parsing of setup.py was failing in several cases,
including when variables were included in the list and comments added
both inline and on their own lines. Packages (see certbot) that have
more than just a basic install_requires list were failing to have any
requirement detected.

This patch detects and removes comments as well as attempts to evaluate
the install_requires list on an item-by-item basis, instead of
literal_eval'ing it at the end. This allows non-literal objects to be
ignored for security reasons.