Add "whatrequires" files for package
The file contains packages that require capabilities of the
specified package using the repoquery --whatrequires command.

This includes the recursive option to query for all packages,
and the archlist=src to specify only SRPMS. To use the
repoquery command a yum.conf file is required; the location
of this is specified in the autospec.conf file. If a path
is not provided, the path is set to the common location of
the yum.conf file in clear.

This file is informational and can be used by other tooling
and package maintainers to identify which packages need to be
rebuilt when this package is updated. It can also be used to
track dependent packages and versions over time by inspecting
git history.

Signed-off-by: Gabi Beyer <gabrielle.n.beyer@intel.com>