Remove remaining VERBOSE make flags
Finish the work of 9bb89743d459ea82f3a517500d46e6d0097cd82e and remove
remaining VERBOSE flags for make. As mentioned in the previous commit
autospec does not parse the verbose output from make, so the flag is not
necessary and causes a considerable slowdown when parsing large build
logs.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>