qmake: add support for LTO builds
qmake has built-in support for LTO (called "LTCG" for "Link-Time Code
Generation" because MSVC had the feature by that name first), so we
don't need to fiddle with the flags by ourselves.

Note that unlike the default support, this does *not* enable fat LTO
archives with Clang and GCC. That means any static libraries added to a
package mut be consumed by the very same compiler (including
version). But since Autospec doesn't package static libraries by
default, that should be no problem.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>