Enable auto exclude for library compat packages
When making a library compatibility package, the only files that are
needed are the library binary and versioned symlinks to the
library. All the other files are to be excluded as they might conflict
with the files in the main package.

To that end, add a setting in options.conf for a package to signify it
is a compatibility library and do matching for files that meet the
above requirements and exclude all others.