Try to autodetect python version by classifiers
If a python package is python3-only and has a setup.py,
"distutils23" will be used as build pattern even if the project is
explicitely specified as python 2 or python 3 only.

This change tries to find if either of these are the case by
looking at the classifiers in the setup.py file for the following
string:

"Programming Language :: Python :: 3 :: Only"
"Programming Language :: Python :: 2 :: Only"

The default is still "distutils23" unless either of these strings
are found.  If they are "distutils2" or "distutils3" will be used
corresponding to the version in the classifier.

Signed-off-by: Kevron Rees <kevron.m.rees@intel.com>