Make git URI configuration optional
Fix #148

Previous to this commit, if the autospec.conf `git` configuration value
is unset, autospec would not instantiate a git repo and not commit any
changes.

However, because the purpose of that config value is to define a pattern
for the git remote URI, it should be completely optional. Users may not
be pushing their repo to a remote location, or the remote repo name may
not match the local repo name (e.g. with Github not allowing certain
characters to appear to repo names).

With this change, autospec will now always commit changes unless
--skip-git is passed on the command line.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>