Move srpms and debuginfo into their own folder
Debug info rpms are not added to bundles, and adding source RPMs will
make mixer crash because it looks for files that will not exist. Break
up the RPMs so it is more obvious which RPMs should be copied over for
building a mix, and which ones are useful meta-data.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>