Add look aside directory for failed patterns
Enable autospec to use an alternate directory for failed pattern
files. These files are optional and are intended to suppliment the
in-repo version of the failed pattern files. The alternate directory
version will supercede the in-repo version of a specific failed
pattern when a conflict between the in-repo version and the
alternative directory file exists.