Start printing patterns that are not detected
We've had bugs sneak into the distro because patterns in upstream
packages subtly changed. This change makes autospec print all patterns
it tries to resolve but is unable to.

As part of this we will need to ignore large list of "autoconf useless
gunk" which is now put in a separate config file.