Fixups for buildreq cache
Only add content to the cache if it hasn't been added before (this
will prevent things that are always added from creeping into the cache
when they are detected to be added as well). Also don't add a newline
when appending content, just an empty string to allow new content to
be put on its own line without adding an empty line between new and
old cache content.