Checking return code after download
Function really_download in tarball.py does not handle http server
codes that can be interpreted as errors i.e. 404 501. When there is no
failure returned in a 404 response the 404 page contents are saved as
the tarball. This change adds code to test that the response from the
server is 200. Any redirection is handled by pycurl internaly and the
code that is checked is the code that the server with the response
returns to autospec.

Signed-off-by: Alex Jaramillo <alex.v.jaramillo@intel.com>