Set test pattern based on build pattern
Instead of selecting tests based on the files source tree, first
check the build pattern being applied to decide what test pattern to
run.

Also remove ruby test patterns as they are not currently used and
need to be reworked to be inline with the rest of the code base.