Check http return code from license server
This change verifies that the HTTP returning code from license server is
a 200 otherwise any code returned by the server will be accepted as a
successful response and the returned error page will be interpreted as a
valid license hash.

Additonally this change narrows the exception emitted by pycurl perform
method from Exception to pycurl.error.

Signed-off-by: Alex Jaramillo <alex.v.jaramillo@intel.com>