Implement mock root.log error detection
Occasionally, packages will add new dependencies that are not yet
available in the repos, or autospec may add an invalid build dependency
for the next build round, etc.

To report these error conditions in a more meaningful way, scan the
root.log after a build whenever mock exits with a non-zero code. If mock
logged a missing dependency (or multiple missing dependencies), report
that information to the user. Otherwise, consider the root.log to be
"clean", and continue by scanning the build.log to determine how to
proceed.

Also, add a new unit test for the root.log parsing function.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>