Add support for multiple destinations per archive
For incredibly complex uses of git submodules, for example, a single
archive may be appropriate to unpack in multiple destinations. First,
track pairs of archives and destinations, second, copy (instead of move)
from the tree where an archive is unpacked into the destination, so that
it can be used again. Order of archives was already respected.